requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
cffi==1.15.1
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
et-xmlfile==1.1.0
greenlet==2.0.1
idna==3.4
multidict==6.0.2
numpy==1.23.4
openpyxl==3.0.7
pandas==1.5.1
psycopg2==2.9.5
pycparser==2.21
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
typing_extensions==4.4.0
yarl==1.8.1
